Python/requirements.txt
google-cloud-pubsub
cryptography
pyjwt
paho-mqtt
